+++ internal/persistence/local/dag_store.go
@@ -13,7 +13,7 @@ import (
 	"github.com/dagu-dev/dagu/internal/persistence/filecache"
 
 	"github.com/dagu-dev/dagu/internal/dag"
-	"github.com/dagu-dev/dagu/internal/grep"
+	"github.com/dagu-dev/dagu/internal/persistence/grep"
 	"github.com/dagu-dev/dagu/internal/util"
 )
